refine

+++ tests/prepare.sh
@@ -75,49 +75,56 @@ elif [ ${MODE} = "infer" ] || [ ${MODE} = "cpp_infer" ];then
 fi
 
 if [ ${MODE} = "cpp_infer" ];then
-    echo "################### build opencv ###################"
     cd deploy/cpp_infer
-    rm -rf 3.4.7.tar.gz opencv-3.4.7/
-    wget https://github.com/opencv/opencv/archive/3.4.7.tar.gz
-    tar -xf 3.4.7.tar.gz
+    use_opencv=$(func_parser_value "${lines[52]}")
+    if [ ${use_opencv} = "True" ]; then
+        echo "################### build opencv ###################"
+        rm -rf 3.4.7.tar.gz opencv-3.4.7/
+        wget https://github.com/opencv/opencv/archive/3.4.7.tar.gz
+        tar -xf 3.4.7.tar.gz
 
-    cd opencv-3.4.7/
-    install_path=$(pwd)/opencv-3.4.7/opencv3
+        cd opencv-3.4.7/
+        install_path=$(pwd)/opencv-3.4.7/opencv3
 
-    rm -rf build
-    mkdir build
-    cd build
+        rm -rf build
+        mkdir build
+        cd build
 
-    cmake .. \
-        -DCMAKE_INSTALL_PREFIX=${install_path} \
-        -DCMAKE_BUILD_TYPE=Release \
-        -DBUILD_SHARED_LIBS=OFF \
-        -DWITH_IPP=OFF \
-        -DBUILD_IPP_IW=OFF \
-        -DWITH_LAPACK=OFF \
-        -DWITH_EIGEN=OFF \
-        -DCMAKE_INSTALL_LIBDIR=lib64 \
-        -DWITH_ZLIB=ON \
-        -DBUILD_ZLIB=ON \
-        -DWITH_JPEG=ON \
-        -DBUILD_JPEG=ON \
-        -DWITH_PNG=ON \
-        -DBUILD_PNG=ON \
-        -DWITH_TIFF=ON \
-        -DBUILD_TIFF=ON
+        cmake .. \
+            -DCMAKE_INSTALL_PREFIX=${install_path} \
+            -DCMAKE_BUILD_TYPE=Release \
+            -DBUILD_SHARED_LIBS=OFF \
+            -DWITH_IPP=OFF \
+            -DBUILD_IPP_IW=OFF \
+            -DWITH_LAPACK=OFF \
+            -DWITH_EIGEN=OFF \
+            -DCMAKE_INSTALL_LIBDIR=lib64 \
+            -DWITH_ZLIB=ON \
+            -DBUILD_ZLIB=ON \
+            -DWITH_JPEG=ON \
+            -DBUILD_JPEG=ON \
+            -DWITH_PNG=ON \
+            -DBUILD_PNG=ON \
+            -DWITH_TIFF=ON \
+            -DBUILD_TIFF=ON
 
-    make -j
-    make install
-    cd ../
-    echo "################### build opencv finished ###################"
+        make -j
+        make install
+        cd ../
+        echo "################### build opencv finished ###################"
+    fi
 
 
     echo "################### build PaddleOCR demo ####################"
-    OPENCV_DIR=$(pwd)/opencv-3.4.7/opencv3/
+    if [ ${use_opencv} = "True" ]; then
+        OPENCV_DIR=$(pwd)/opencv-3.4.7/opencv3/
+    else
+        OPENCV_DIR=''
+    fi
     LIB_DIR=$(pwd)/Paddle/build/paddle_inference_install_dir/
-    CUDA_LIB_DIR=/usr/local/cuda/lib64/
-    CUDNN_LIB_DIR=/usr/lib/x86_64-linux-gnu/
-
+    CUDA_LIB_DIR=$(dirname `find /usr -name libcudart.so`)
+    CUDNN_LIB_DIR=$(dirname `find /usr -name libcudnn.so`)
+    
     BUILD_DIR=build
     rm -rf ${BUILD_DIR}
     mkdir ${BUILD_DIR}
